+++ resolwe/test/utils.py
@@ -92,6 +92,9 @@ def with_custom_executor(wrapped=None, **custom_executor_settings):
                 # Re-run engine discovery as the settings have changed.
                 manager.discover_engines()
 
+                # Re-run the post_register_hook
+                manager.get_executor().post_register_hook()
+
                 # Run the actual unit test method.
                 return wrapped_method(*args, **kwargs)
         finally: